requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.3.1            # via flask-migrate
aniso8601==8.0.0          # via flask-restful
apscheduler==3.6.3        # via -r requirements.in
authlib==0.12.1           # via -r requirements.in, loginpass
blinker==1.4              # via sentry-sdk
cachetools==3.1.1         # via google-auth
certifi==2019.11.28       # via requests, sentry-sdk
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via -r requirements.in, flask
cryptography==2.8         # via authlib
flask-caching==1.7.2      # via -r requirements.in
flask-cors==3.0.8         # via -r requirements.in
flask-marshmallow==0.11.0  # via -r requirements.in
flask-migrate==2.5.2      # via -r requirements.in
flask-restful==0.3.7      # via -r requirements.in
flask-sqlalchemy==2.4.1   # via -r requirements.in, flask-migrate
flask-wtf==0.14.2         # via -r requirements.in
flask==1.1.1              # via -r requirements.in, flask-caching, flask-cors, flask-marshmallow, flask-migrate, flask-restful, flask-sqlalchemy, flask-wtf, sentry-sdk
google-api-python-client==1.7.11  # via -r requirements.in
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.7.2        # via google-api-python-client, google-auth-httplib2
httplib2==0.14.0          # via google-api-python-client, google-auth-httplib2
humanize==0.5.1           # via -r requirements.in
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.3            # via flask
loginpass==0.4            # via -r requirements.in
mako==1.1.0               # via alembic
markdown==3.1.1           # via -r requirements.in
markupsafe==1.1.1         # via jinja2, mako
marshmallow-sqlalchemy==0.22.3  # via -r requirements.in
marshmallow==3.5.1        # via flask-marshmallow, marshmallow-sqlalchemy
musicbrainzngs==0.7.1     # via -r requirements.in
netaddr==0.7.19           # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
pyasn1-modules==0.2.7     # via -r requirements.in, google-auth
pyasn1==0.4.8             # via -r requirements.in, pyasn1-modules, rsa
pycparser==2.19           # via cffi
pymysql==0.9.3            # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in, alembic
python-editor==1.0.4      # via -r requirements.in, alembic
pytz==2019.3              # via -r requirements.in, apscheduler, flask-restful, tzlocal
redis==3.3.11             # via -r requirements.in
requests==2.22.0          # via -r requirements.in, loginpass
rsa==4.0                  # via google-auth
sentry-sdk[flask]==0.13.5  # via -r requirements.in
six==1.13.0               # via apscheduler, cryptography, flask-cors, flask-marshmallow, flask-restful, google-api-python-client, google-auth, python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.36.0  # via -r requirements.in
sqlalchemy==1.3.11        # via -r requirements.in, alembic, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-utils
tzlocal==2.0.0            # via apscheduler
unidecode==1.1.1          # via -r requirements.in
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.7           # via requests, sentry-sdk
werkzeug==0.16.0          # via flask
wtforms==2.2.1            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
